+++ contrib/anthropic/README.md
