+++ examples/go.mod
@@ -4,8 +4,8 @@ go 1.24.6
 
 replace (
 	github.com/go-kratos/blades => ../
-	github.com/go-kratos/blades/contrib/claude => ../contrib/claude
-	github.com/go-kratos/blades/contrib/gemini => ../contrib/gemini
+	github.com/go-kratos/blades/contrib/anthropic => ../contrib/anthropic
+	github.com/go-kratos/blades/contrib/google => ../contrib/google
 	github.com/go-kratos/blades/contrib/mcp => ../contrib/mcp
 	github.com/go-kratos/blades/contrib/openai => ../contrib/openai
 	github.com/go-kratos/blades/contrib/otel => ../contrib/otel
@@ -13,11 +13,10 @@ replace (
 
 require (
 	github.com/go-kratos/blades v0.0.0
-	github.com/go-kratos/blades/contrib/gemini v0.0.0-20251106103709-242709515a73
+	github.com/go-kratos/blades/contrib/google v0.0.0-00010101000000-000000000000
 	github.com/go-kratos/blades/contrib/mcp v0.0.0-20251106103709-242709515a73
 	github.com/go-kratos/blades/contrib/openai v0.0.0-20251106103709-242709515a73
 	github.com/go-kratos/blades/contrib/otel v0.0.0-20251106103709-242709515a73
-	github.com/go-kratos/kit v0.0.0-20251107074530-fa502981ad29
 	github.com/google/jsonschema-go v0.3.0
 	go.opentelemetry.io/otel v1.38.0
 	go.opentelemetry.io/otel/exporters/stdout/stdouttrace v1.38.0
@@ -30,6 +29,7 @@ require (
 	cloud.google.com/go/auth v0.17.0 // indirect
 	cloud.google.com/go/compute/metadata v0.9.0 // indirect
 	github.com/felixge/httpsnoop v1.0.4 // indirect
+	github.com/go-kratos/kit v0.0.0-20251107074530-fa502981ad29 // indirect
 	github.com/go-logr/logr v1.4.3 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/google/go-cmp v0.7.0 // indirect
